// src/lib/api/generated/model/publicContentSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export interface PublicContentSchema {
  audience: string;
  body_html: string;
  category?: string | null;
  content_type?: string | null;
  external_url?: string | null;
  featured?: boolean | null;
  id: string;
  publish_end?: string | null;
  publish_start?: string | null;
  published_at?: string | null;
  resource_type?: string | null;
  slug: string;
  status: string;
  summary: string;
  target_scope_mode: string;
  title: string;
}